demo4_ai_platform/requirements.txt
daft[lance]>=0.4.0
lancedb>=0.15.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pyyaml>=6.0
torch>=2.0.0
